Moved the compiler cmds
