+++ conf/script/src/build_system/cmd/compiler/host/get_info/__init__.py
@@ -1,0 +1,1 @@
+import build_system.cmd.compiler.host.get_info.cli
