+++ conf/script/src/build_system/cmd/compiler/host/get_info/version/__init__.py
@@ -1,0 +1,1 @@
+from build_system.cmd.compiler.host.get_info.version import compiler
