+++ conf/script/src/build_system/cmd/compiler/host/get_info/version/gnu/__init__.py
@@ -1,0 +1,1 @@
+from build_system.cmd.compiler.host.get_info.version.gnu.fetch_gnu_compiler_version import fetch
