+++ conf/script/src/build_system/cmd/compiler/host/get_info/version/gnu/fetch_gnu_compiler_version.py
@@ -2,12 +2,11 @@ import subprocess
 from pathlib import Path
 from typing import AnyStr
 
-from build_system.compiler.host.get_info import version
 from build_system.compiler.version import CompilerVersion
 
 
 def _fetch_raw(compiler: Path) -> AnyStr:
-    version.compiler.assure_path_integrity(compiler)
+    build_system.cmd.compiler.host.get_info.version.compiler.assure_path_integrity(compiler)
 
     result: subprocess.CompletedProcess = subprocess.run(
         [compiler, '-dumpversion'], capture_output=True, text=True, check=True
@@ -17,4 +16,4 @@ def _fetch_raw(compiler: Path) -> AnyStr:
 
 
 def fetch(compiler: Path) -> CompilerVersion:
-    return version.compiler.fetch(compiler, _fetch_raw)
+    return build_system.cmd.compiler.host.get_info.version.compiler.fetch(compiler, _fetch_raw)
