+++ conf/script/src/build_system/cmd/compiler/host/get_info/version/msvc/__init__.py
@@ -1,0 +1,1 @@
+from build_system.cmd.compiler.host.get_info.version.msvc.fetch_msvc_version import fetch
