+++ conf/script/src/build_system/cmd/compiler/host/get_info/version/msvc/__main__.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+
+import utils.cli.main
+
+
+def main():
+    build_system.cmd.compiler.host.get_info.version.msvc.cli.fetch_version()
+
+
+utils.cli.main.wrap_main(main)
