--- conf/script/src/build_system/compiler/host/get_info/__init__.py
@@ -1,1 +1,0 @@
-import build_system.compiler.host.get_info.cli
