--- conf/script/src/build_system/compiler/host/get_info/location/msvc/__init__.py
@@ -1,1 +1,0 @@
-from build_system.compiler.host.get_info.location.msvc.installation_path import find
