--- conf/script/src/build_system/compiler/host/get_info/version/__init__.py
@@ -1,1 +1,0 @@
-from build_system.compiler.host.get_info.version import compiler
