--- conf/script/src/build_system/compiler/host/get_info/version/gnu/__init__.py
@@ -1,1 +1,0 @@
-from build_system.compiler.host.get_info.version.gnu.fetch_gnu_compiler_version import fetch
