--- conf/script/src/build_system/compiler/host/get_info/version/msvc/__init__.py
@@ -1,1 +1,0 @@
-from build_system.compiler.host.get_info.version.msvc.fetch_msvc_version import fetch
