+++ conf/script/src/test/test_build_system/test_compiler/test_host/test_get_info/test_location/test_msvc/test_installation_path.py
@@ -5,7 +5,7 @@ from unittest import mock
 from unittest.mock import MagicMock
 
 import utils.cli.main
-from build_system.compiler.host.get_info.location.msvc import installation_path
+from build_system.cmd.compiler.host.get_info.location.msvc import installation_path
 
 
 class TestFind(unittest.TestCase):
